rename: rename Member class to MemberEntity to clarify that the class is only used as a payload to contain values to be inserted into user table
