+++ src/test/java/com/find/doongji/history/service/HistoryServiceTest.java
@@ -4,8 +4,7 @@ import com.find.doongji.auth.enums.Role;
 import com.find.doongji.history.payload.request.HistoryRequest;
 import com.find.doongji.history.payload.response.HistoryResponse;
 import com.find.doongji.history.repository.HistoryRepository;
-import com.find.doongji.member.payload.request.SignUpRequest;
-import com.find.doongji.member.payload.response.Member;
+import com.find.doongji.member.payload.request.MemberEntity;
 import com.find.doongji.member.repository.MemberRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
@@ -35,14 +34,14 @@ public class HistoryServiceTest {
     @BeforeEach
     void setUp() {
         String username = "testuser";
-        Member member = Member.builder()
+        MemberEntity memberEntity = MemberEntity.builder()
                 .username(username)
                 .email("test@gmail.com")
                 .name("Test User")
-                .role(Role.ROLE_USER)
+                .role(Role.ROLE_USER.getKey())
                 .password("password")
                 .build();
-        memberRepository.insertUser(member);
+        memberRepository.insertMember(memberEntity);
     }
 
     @Test
